Add a blacklist for devices not be scanned for UBI
By attaching a UBI device to a MTD partition that is not
thought to be used with UBI damages will damage the
current content. UBI tries to fix issues that do not exist
because the device is not used at all with UBI.

Adding a blacklist, set at compile time with a
configuration option, let to exclude a list
of devices when devices are searched for UBI
volumes.

Signed-off-by: Stefano Babic <sbabic@denx.de>