Added handler for raw flash
The handler for raw flash was not working and
there is no support for NAND flash. In fact,
writing to a NAND in raw mode (without UBI layer)
requires to take care of bad blocks.
Code was taken from mtd-utils (nandwrite) to
work correctly with NAND. Most of options
in nandwrite are dropped here because they have
no meaning for swupdate.

Signed-off-by: Stefano Babic <sbabic@denx.de>