Use revision strings instead of major.minor
Boards have often a more complicated schema
to identify itself with a revision number. The
simple way with two integers for a major.minor
revision is not suitable for that purposes.
Replace them with a string, that identify the
revision for a board. The Hardware compatibility
file must have the format:
	<target name>	<target revision string>

Only one entry is allowed and read by the parser.
Further entries are ignored.

Signed-off-by: Stefano Babic <sbabic@denx.de>