Handling configuration differences
Add support for having a single image containing
the software for multiple devices. Support is
added to the libconfig parser.

The resulting image is then realized as the concatenation
of multiple images, one for each single device.
Files inside the whole image must have unique filenames.

sw-description is extended to have an entry for each target.

software =
{
	version = "0.1.0";

	target-1 = {
		images: (
			{
				...
			}
		);
	};

	target-2 = {
		images: (
			{
				...
			}
		);
	};
}

Previous format is still supported. If there is not an entry
of type software.<board name>.<field>, the parser falls back
to software.<field>

Signed-off-by: Stefano Babic <sbabic@denx.de>