Drop U-Boot specifig CONFIG_ switch
The functions for managing the U-Boot's environment
are taken from U-Boot project and use several switches
that are defined when U-Boot is configured for a specific board.

swupdate does not use them, and there is not
such board configuration. The generated default environment
is then wrong and should not be used at all.
swupdate always need that the environment is described
by fw_env.config and other cases are dropped from code.

Signed-off-by: Stefano Babic <sbabic@denx.de>