Make check for adjusting UBI volumes less strong
A volume is repartitioned if the size is not equal
to the value set in the sw-description. However,
a volume is always created as integer number of LEBs,
and this means that the volume size is rounded up
to the next multiple number of LEB by the kernel.
This makes that a volume is always repartitioned
if in sw-description is not set as multiple of LEB size.

Make the check depending on the LEB size, and do not
repartition the volume is number of required LEBs
does not change.

Signed-off-by: Stefano Babic <sbabic@denx.de>