Move in progress flag before installing
the recovery_status U-Boot variable is used
to signal to the bootloader the start of
a "swupdate transaction". It is reset
only after a successful update.
If a user tries to install a wrong image
before something is really installed on the target,
the variable is set and after a reboot the
target is still in swupdate, even if a valid
software (old version) is still untouched.

Move the set of the variable after all checks
and before installing on the target - if
nothing is installed, the boot loader
can start the old software again.

Signed-off-by: Stefano Babic <sbabic@denx.de>