Fix image filename with international characters
Installing from a file, the image is not found
if its name uses a character set that is not supported.
This happens with japanese and several other
languages. The filename is passed from the command line,
and international support must be added to
get the correct filename.

Instead of this, let that swupdate search for
a well defined pattern into a directory. If the
passed image file is a pattern, such as :
	swupdate -i "/media/usb/*.swu"

swupdate searches for all entries in the directory
and takes the first of them matching the pattern.

Signed-off-by: Stefano Babic <sbabic@denx.de>