Add JSON parser
Current default parser is based on the libconfig
library. Its syntax is near to JSON, but it is
not a standard JSON. Add a JSON parser based on
json-c library. Syntax is adjusted to be
conform to JSON.

Reason for this is to allow external software
to generate automatically the sw-description file
with standard tools. Because JSON is large available
on all architectures and in most programming languages,
it is easier to write such as software as by using libconfig.

Signed-off-by: Stefano Babic <sbabic@denx.de>