Do not change order in RECOVERY_STATUS enum
RECOVERY_STATUS is used to inform the status of an update.
This can be used by automatic software or by external GUI
to start an update vie the interface with the integrated
webserver, and changing the order will break any automated
software.

Commit 44a57b5 changed the order and breaks the behavior.
The patch sets the order back and adds comments to explain
why changing the order is not allowed.

Signed-off-by: Stefano Babic <sbabic@denx.de>