Fix working with IE11
IE11 added compatibility to the other browsers
regarding Ajax requests. However, IE does pretty aggressive
caching, and polling from the webserver to get
status of the current update is not working.

Disable cache in the Ajax connection, letting the
status to be correctly displayed as on other
browsers (Chromium, Firefox).

Signed-off-by: Stefano Babic <sbabic@denx.de>