Add IPC library to use swupdate as installer daemon
Instead of using the integrated Webserver,
the new software can be streamed by an external
application. The main application has already
its way to communicate with the world, and in
some cases is not desired to have a second way
to get into the target, like using the webserver.
This is specially helpful when swupdate is not
used as rescue system, but in the dual-copy
option. swupdate is started as separate process
and it is controlled by the main application.

Signed-off-by: Stefano Babic <sbabic@denx.de>