Provide local version of sys/queue.h
Not all systems provide a sys/queue.h implementation. For example the the musl
C library does not and will not include it [1].

sys/queue.h first appeared in 4.4BSD and is used in all 4.4BSD variants. Note
that glibc' sys/queue.h is a patched version of this 4.4BDS header to add
STAILQ_CONCAT and TAILQ_CONCAT.

There exists several implementations of sys/queue.h. We use the version from the
NetBSD project [2] and provide it as a local header file "bsdqueue.h"

[1] http://wiki.musl-libc.org/wiki/FAQ#Q:_why_is_sys.2Fqueue.h_not_included_.3F
[2] http://cvsweb.netbsd.org/bsdweb.cgi/src/sys/sys/queue.h?rev=1.68

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>