scripts: replace kbuild/kconfig files
The previous kbuild/kconfig infrastructure used by swupdate is paritally
broken (xconfig, gconfig) and several features are missing, e.g. passing
parameters through the environment with the env option or the make target
savedefconfig.

We are using a kbuild template based on the Linux kernel 4.2.0-rc1 [1].
The following patches will adapt swupdate to use this template for building.

[1]
https://github.com/joerg-krause/kbuild_skeleton

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>