Kbuild: fix build issue with some toolchains
Some toolchains used in Buildroot (x86-64 and mips64) [1] have a problem to
compile a set of object files into one object file as done for the
`builtin-target`. This targets collects all object files from each subdirectory
declared with `objs-dir`: core and handlers.

For this target the default emulation of ld is not the correct one resulting in
a relocation error:

  /usr/bin/i686-pc-linux-gnu-ld: Relocatable linking with relocations from
  format elf64-x86-64 (core/swupdate.o) to format elf32-i386 (core/built-in.o)
  is not supported.

As there is no need to have `built-in.o` files for core and handlers we can
easily allow swupdate to be built with these toolchains by compiling the source
files in core and handlers with the `lib-target` target as it is done for the
other subdirectories.

The `lib-target` compiles a set of object files into an archive file. With this
target we avoid using the ld linker and use the ar archiver instead.

[1] http://autobuild.buildroot.net/results/d9e/d9e8298898da6355f6b46a032f951b4ca196d103/build-end.log

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>