UBI: check for UBI before attaching
Inconditionally attaching a MTD device to check
UBI volumes has several drawbacks, specially when
swupdate runs in double-copy mode. In fact, it is
not unusual that just a single MTD device is attached,
and the copies are different volumes inside the same
container. ubi_attach() cannot be successful,
because the MTD device is already attached and cannot be
detached because contains the current root filesystem.
This constraints to have different MTDs, one for each
software copy.

Patch removes this limitation: UBI containers are scanned
before attaching a device, and a new device is just
attached if it was not before. Multiple copies
can be then stored in the same UBI container.

Signed-off-by: Stefano Babic <sbabic@denx.de>