handlers: add a handler to write 1bit hamming OOB
Writes a loader binary to a NAND flash memory device and calculates
1-bit Hamming ECC codes to fill the MTD's out-of-band (oob) area
independently of the ECC technique implemented on the NAND driver.
This is a workaround required for TI ARM ROM code to load the SPL.

Code is based on [1] 0469baaa7def, "writeloader: avoid -lm"

[1] https://github.com/martinezjavier/writeloader

Tested on a TI OMAP AM3517.

Signed-off-by: Jeroen Hofstee <jeroen@myspectrum.nl>
Signed-off-by: Stefano Babic <sbabic@denx.de>