Resume a download if it was interrupted
Allow to resume a download if it was interrupted. This
uses the libcurl API to reactivate it. The download
is marked as broken if there is an error or if
a very low number of bytes is transferred in the last
minutes. Values are fixed and set with the defines:

	DL_LOWSPEED_BYTES
	DL_LOWSPEED_TIME

Default value is 8bytes/sec in the last 5 minutes. This should
allow even very low-bandwidth connections.

Signed-off-by: Stefano Babic <sbabic@denx.de>