Conditional update from installed version
Avoid to install an image if the same is already installed.
This can be used to reduce risk of brick the target
in case an image is unique on the system and the update
cannot be safe. Typical case is the bootloader.
Checking the installed version, an image is installed
only if it differs from the version present on
the system.

Signed-off-by: Stefano Babic <sbabic@denx.de>