Drop obsolete U-Boot functions to set environment
fw_env.c in swupdate is obsolete - U-Boot functions
are exported by the U-Boot project.
Copy in swupdate is old and does not work with many
devices, for example with SD/eMMC.

If you get a link error, check to have libubootenv
in your build. If not, the library is the copy
of tools/env/lib.a from U-Boot.

Signed-off-by: Stefano Babic <sbabic@denx.de>