Makefile: fix expansion of CROSS_COMPILE
Commit 7780a9e76ac0e01d152bb83850bbed67a403f8c5 breaks cross-compilation if
CONFIG_CROSS_COMPILE is set through Kconfig.

The problem is that the value passed to eval is expanded twice; first by the
eval function, then the results of that expansion are expanded again when they
are parsed as makefile syntax [1]. Variables that should to be expanded after
evaluation must be escaped.

[1] https://www.gnu.org/software/make/manual/html_node/Eval-Function.html

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>