Fix licensing issue due to cpio
The main license of the project is GPLv2+ (as stated in the COPYING file and
confirmed by most of the files). But the file "include/cpiohdr.h" is
licensed under the GPLv3+. As a result the effective license for the
whole project to resolve the licensing conflict is GPLv3+.

However, the file is not the original one from the cpio tool:
swupdate just requires a couple of structures to be defined
to inspect the cpio archive. To solve the issue,
the file is rewritten and licensed under GPLv2.

Signed-off-by: Stefano Babic <sbabic@denx.de>
CC: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>