Suricatta framework
Suricatta is -- like the existing mongoose -- a daemon mode of SWUpdate,
hence the name suricatta (engl. meerkat) as it belongs to the mongoose
family. Suricatta regularly polls a remote server for updates and
installs them.

This initial commit introduces the foundational suricatta framework for
concrete server and channel implementations to base on.

Signed-off-by: Christian Storm <christian.storm@tngtech.com>
Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>