Enhance output when searching for LUA handlers
The LUA interpreter shows an error if support for
handlers written in LUA is activated and the file
swupdate_handlers.lua is not found. This is not an error,
simply no additional handlers are registered into the
system. Try to explain this, and output that no
LUA handlers are found.

Signed-off-by: Stefano Babic <sbabic@denx.de>