Warn if sw-description key length > MAX || 0
Using an sw-description file containing
    ...
    uboot: ({
            name  = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
            value = "0123456789"
    }
    ...
results in
    ABCDEFGHIJKLMNOP0123456789=0123456789
being written to the U-Boot environment instead of the expected result
    ABCDEFGHIJKLMNOPQRSTUVWXYZ=0123456789

The key "ABCDEFGHIJKLMNOP0123456789" is the concatenation of
* the sw-description's "name" value cropped to UBOOT_VAR_LENGTH and
* sw-description's "value" value
since reading from struct uboot_var's .varname lacks \0 termination so
that the key's string extends to the next \0 terminator, being the
one of uboot_var.value.

This patch adds a \0 terminator in this case and issues a warning.
Additionally, a warning is issued if the key's length is 0.

Signed-off-by: Christian Storm <christian.storm@tngtech.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>