Introducing progress bar interface
There is not a clean interface to get a progress
status from swupdate. The only way is to require the status
polling the ctrl interface, and reading the string
sent to the interface.

Add a new interface via UDS just to inform another
process how the installation is going on.
The client needs to connect and it will be notified
at each change in the installation process.
This simplifies the realisation of a GUI,
displaying which image is installed.

Signed-off-by: Stefano Babic <sbabic@denx.de>