Rename swupdate_dgst_cleanup
swupdate_dgst_cleanup is not related to the whole process,
but it is called after verifying the single hash of
an image. It is better to call it swupdate_HASH_cleanup
because just the openSSL CTX for the single hash verification
is freed, not the structure of the whole process.

Signed-off-by: Stefano Babic <sbabic@denx.de>