Replace copyfile function for handlers with easier copyimage
copyfile() takes care of a lot of parameters. Handlers
do not need all of them and the interface for the handler
appears confuse.

Simplify the interface for the handler introducing a simpler
copyimage(int , struct img_type *) function. All parameters
for extracting a copying a single image can be derived
from the img_type struct.

Signed-off-by: Stefano Babic <sbabic@denx.de>