downloader: properly terminate after maximum retries
If the download fails to complete after the maximum number of retries,
the installer thread will be stuck for ever waiting for data.  This
causes ipc_wait_for_complete() to never return, and the swupdate
process never finishes.

Fix this by calling ipc_wait_for_complete() only if the download
was successful, otherwise simply terminating the process with
ipc_end().

Signed-off-by: Mans Rullgard <mans@mansr.com>