copyfile: substitute fdout with arbitrary data
A callback for writing data could write somewhere else
instead of into a file descriptor. Pass an arbitrary
pointer to the copy- functions. If callback is not set,
the pointer is set to the address of a valid file
descriptor. When callback is set, the pointer must be
set to a useful data for the callback.

Signed-off-by: Stefano Babic <sbabic@denx.de>