Allow to run swupdate just to check the image.
During build process, it makes sense to check the swu
image before installing on target.

SWUpdate can run on the host, and checks if
the sw-description is correct and all
required images are part of the compound image.

The check is activated with the parameter:
    swupdate --check --image <SWU to check>

In compare to the --dry-run option suggested by Georg McCollister,
the --check option will just ensure that the files expected by
swupdate are present.

Signed-off-by: Denis Osterland <Denis.Osterland@diehl.com>