suricatta: drop duplicate parameters for loglevel and verbose
loglevel and verbose are already set in the main module and it
is not necessary to set it again into the suricatta module.
This has also the drawback that a global set is overloaded
by suricatta.

In fact, this was allowed:

	swupdate -l 1 -u "-l 5"

logleve is passed twice, and suricatta's value is taken
as global value (loglevel is valid globally).

Signed-off-by: Stefano Babic <sbabic@denx.de>