Support sha256 even if SIGNED_IMAGES is not set
sha256 is working only if CONFIG_SIGNED_IMAGES  is set.
If a sha256 is added to an artifact, SWUpdate stops
with an error because the interface to the hash
functions are disabled in sslapi.h.

Split hashing and verification adding CONFIG_HASH_VERIFY,
that is automatically set when CONFIG_SIGNED_IMAGES
is enabled.

Signed-off-by: Stefano Babic <sbabic@denx.de>