progress: don't die with SIGPIPE on disconnected socket
A write() on a disconnected socket returns EPIPE and sends the SIGPIPE
signal to the running thread. The latter is not handled by swupdate
which is why it dies in this case.
Use signal() with the MSG_NOSIGNAL flag to survive on trying to send
a message to a disconnected socket.

Signed-off-by: Christian Storm <christian.storm@tngtech.com>
Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>