remote handler: get timeout value from partner
The remote installer can need more time for
installing, for example if it has to erase some flash.
Let the remote to comunicate the time it needs
before sending the answer. This is part of
the "ACK" answer, concatenated with the ":"
separator.
The remote can send then a packet like:
	ACK:<timeout value>

Signed-off-by: Stefano Babic <sbabic@denx.de>