Move code to introduce process control
Security reasons lead that all processes
that communicates via network should not run as root.
SWUpdate, even if does not run as root, must have
high privileges to be able to change the software,
that means the access of most hardware devices must be
guaranteed.
It is desirable to split SWUpdate in a multi-process
application, where network processes can be downgraded
to another user.

Add functions for moving some
parts of SWUpdate into separate
processes.

Signed-off-by: Stefano Babic <sbabic@denx.de>