suricatta: fix and improve proxy support
Allow the specification of a proxy to be used at run-time rather than
at compile-time (via SURICATTA_HAWKBIT_PROXY).

The new command-line option '-y' can optionally be augmented with
a proxy URL. If no proxy URL is given on the command-line, the
environment variables http_proxy and all_proxy are considered. If no
valid command-line proxy configuration is found, an error is issued.

The according config file option is named 'proxy'.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>