parser: report error if no images etc found
If the requested software set, as provided with the -e flag, does not
exist in the sw-description, swupdate silently does nothing and returns
success.  This behaviour is confusing in case a typo or similar causes
a bad set to be requested.

Fix this by making the parser report an error if it finds nothing
matching the requested software set.

Signed-off-by: Mans Rullgard <mans@mansr.com>