suricatta: do not allow empty variable for state
SURICATTA_STATE_UBOOT can be set to an empty
string, but this generates an entry into the
U-Boot's environment with a value and without
a name.
When the board reboots, U-Boot thinks that
the environment is corrupted and board does
not boot anymore. Check if the state variable
is set, and if not reports an error and avoid
to write into the environment.

Signed-off-by: Stefano Babic <sbabic@denx.de>