Segfault using a progress client
When using a download progress client which
connects and disconnects before SWupdate is finished,
the process segfaults.

This is due to the wrong macro to iterate the
list of connections. The SAFE version uses  a temporary
variable to hold the next element, so that an entry
may be freed or removed from the list.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Jörg Krause <joerg.krause@embedded.rocks>