hawkbit: proxy set even if no entry in configuration file
server_hawkbit calls repeatedly GET_FIELD_STRING,
but this is thought to not change the destination
if no entry in configuration file is found.
Using the same variable without resetting it
makes further variables set as the ones before.

Add a new macro GET_FIELD_STRING_RESET that
additionaly reset the destination string to "".

Signed-off-by: Stefano Babic <sbabic@denx.de>