Change SIGCHLD handler to let work with system()
If a shellscript is executed, the installer calls
system(), that starts a new process and waits
until this process finishes. But SIGCHLD
is catched by a different thread (the main thread) as the installer
thread, and before system() has a chance to do this.

Rework the SIGCHLD handler to avoid to wait for any child,
and calls waitpid just for the processes that must
be monitored.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Jörg Krause <joerg.krause@embedded.rocks>
Tested-by: Jörg Krause <joerg.krause@embedded.rocks>