hawkbit: send configData before going to update
Do not start an update if configData was not
yet sent. First try to send the static data,
and at the next polling cycle try to retrieve
the new software, if any.

Reduce the polling cycle after configData was sent
to check as soon as possible if a new
update is available.

Signed-off-by: Stefano Babic <sbabic@denx.de>