core: post-update command execution
After having installed an update, an external agent (listening
on the progress interface) gets notified and should decide on
further actions to take if any, e.g., shutdown all running
applications and reboot.

In addition to that, a command to be executed via system(3)
may be given by the --postupdate command line option or the
config file.

Signed-off-by: Christian Storm <christian.storm@siemens.com>