tests: silence acceptance tests output
The acceptance tests' output includes parameter error messages from
swupdate as it is the case, e.g., for the suricatta test. This is
confusing at first sight. Hence, silence the output so that only test
progress is visible.

Failing tests can still be easily spotted by running
$ make -j1 V=1 acceptance-tests
and then executing the culprit test manually to see it's output.

Signed-off-by: Christian Storm <christian.storm@siemens.com>