tests: integrate suricatta tests into kbuild
Suricatta's tests have lived "off-tree", resulting in some ongoing
maintenance efforts to keep method signatures for mocks and the like in
sync. Additionally, kconfig options had to be set specifically for the
tests in the tests' Makefile, resulting in a potential divergence of the
tests' kconfig options and a swupdate's kconfig options.

By integrating suricatta's tests into the kbuild infrastructure, both
these problems are addressed and solved.

Signed-off-by: Christian Storm <christian.storm@siemens.com>