Hawkbit: dynamically allocated channels and cancelUpdate
The server uses a static instance for the channel
interface used to communicate with the Hawkbit server.
This has the disadvantage that just one connection at
time is possible and a check for the status on the server
is not possible if a connection is already running.
Such case happens when artifacts are downloaded,
while on the server the cancel was pressed.

The status should be polled as usual to check
if the current istallation must aborted.

Move to an allocated channel and store the pointer
to the channel_t structure inside the server.

Add a callback function for the downloader
to be called after writing the data. The
callback is used to ask the Hawkbit server
if a cancelUpdate was requested.

Signed-off-by: Stefano Babic <sbabic@denx.de>