IPC: introduce ipc for SWUpdate components
Splitting SWUpdate in processes requires to
implement an IPC if they need to communicate
together. First use case is to send back
a feedback to the hawkbit server from an
external process.

SWUpdate will maintain a single socket
to the world. A different message is implemented
and the network daemon routes the messages
to the right SWUpdate's process.

Signed-off-by: Stefano Babic <sbabic@denx.de>