hawkbit: manage IPC and add feedback from outside
Add IPC to allow external processes to send
messages / command. This patch adds also
a "feedback" command. This is useful
if the verification of the installed image
is not done by SWUpdate, but demanded to
an external process. In this case,
suricatta should be started with "-c 6"
(new state, WAIT) and no communication
is started to the Hawkbit server until
getting the result from the IPC.

Signed-off-by: Stefano Babic <sbabic@denx.de>