Remove MAX_MTD_DEVICES limitation
swupdate will refuse to work with any MTD partition whos number
is above MAX_MTD_DEVICES (20). In practise, this means you cannot
update any partition above /dev/mtd20 with the current swupdate.

This patch removes this limitation, as there is no reason for having
it. MAX_MTD_DEVICES was introduced in very early versions of swupdate
when a static array was used to hold MTD partition data; nowadays it
it is dynamically allocated based on number of MTD partitions available
in the system.

If you want to limit the MTD partitions swupdate may operate on, use
the MTD blacklisting feature.

Signed-off-by: Magnus Olsson <magnus@minimum.se>
Acked-by: Stefano Babic <sbabic@denx.de>