json parser: failed to parse compatibility
Passing NULL as key to json_object_object_get_ex()
does not return the current node as supposed.

Parsing the hardware-software compatibiolity, the
key is NULL because it is not required to travers the tree:
the passed entry is already the node to be retrieved.

Check for path in get_field_string_json() and
avoids to search again for the key.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Renaud Barbier <renaud.barbier@abaco.com>