hawkbit: check that filename is a .swu image
The update URL can contain multiple artifacts, because the same hardware can
run different update agent. The issue is raised with a target that move to
SWUpdate from its previous updater, but for compatibility reason the targets
in field maintain the old updater. The same release contains both new and old
format. If the filename is not checked before sending to the installer,
SWUpdate breaks because it is not a CPIO archive.

Check that the file is a ".swu" image and skip all other files.

Signed-off-by: Stefano Babic <sbabic@denx.de>