Terminate child processes if parent dies
The installer process monitors the sub-processes and detects their death, this
is not true the other way round: if the main process dies, sub-processes
remain running.

This will prevent respawning the main process because it in turn will
fail to start sub-processes, as they are already running and holding the
designated network port.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Sami Hartikainen <sami.hartikainen@teleste.com>