notifier: add a notifier to forward to progress
A sub process can have some information that it is not needed by the
installer, but it can be useful for an external process or for everyone
using the progress interface. Add a new notifier that check for such
information and forward them as "EVENT" to the progress interface.

In the progress message, the EVENT is coded in the info field.

Signed-off-by: Stefano Babic <sbabic@denx.de>