hawkbit: report error if no artifacts are installed
If the artifacts[] is empty or does not contain a valid SWU image,
nothing is installed and a successful update is reported. However, this
lead a controlling software to think that a real update was done and can
initiate a reboot, switching working copy for dual-copy option.

This is wrong, and it is better to signal that something went wrong on
the server.

The issue is not easy to reproduce, but it happens in case of corruption
of MongoDB on the server.

Signed-off-by: Stefano Babic <sbabic@denx.de>