Allow to use elapsed certificates
Even if having elapsed certificates should be avoided, it can be a very
severe issue on embedded systems. In gact, a device can be produced and
stocked before to be sold, and in the meantime the certificate can
elapse. And this is even worse, because the device needs surely an
updated software.

Let configure the behaviour: two additional configuration options are
added.

CMS_IGNORE_EXPIRED_CERTIFICATE : a certificate is valid even if elapsed

CMS_IGNORE_CERTIFICATE_PURPOSE: allow to use the certificate even if not
thought for digital sign.

Of course, the strict behaviour is still possible and it is the default
behaviour (certificate not accepted in case of error).

Signed-off-by: Stefano Babic <sbabic@denx.de>