hawkbit: change ipc to exchange JSON message
The IPC to suricatta is a position dependent list of strings, making the
call quite abstruse and be prone to errors. As the module must have full
support for JSON, change the interface for the IPC to get a JSON string
that can be parsed using json-c calls.

The supported fields are status, finished, execution, details

Signed-off-by: Stefano Babic <sbabic@denx.de>