progress: do not drop the RECOVERY_STATUS
The status is not forwarded by the notifier and it is silently dropped.
Let it go through the progress to allow the receiver to take decisions
on base of the state.

The use case is when an Update is cancelled by the hawkbit server. The
update was not successful, and this information must be forward to the
receiver to avoid that an aborted update is considered ok.

Signed-off-by: Stefano Babic <sbabic@denx.de>