hawkbit: report errors to the Hawkbit Server
If an update fails, just a generic error is reported to the server. It
is not possible to check on the server what is happening without having
a look at SWUpdate's output.

This patch intercepts all ERROR messages during an update and collect
them in case of failure. In this case, they are sent to the backend and
they can be displayed with Hawkbit's GUI in the history window.

Signed-off-by: Stefano Babic <sbabic@denx.de>