copyfile: add new property to handle offset for raw handler
This new property allows the user to specify a specific offset from a
raw device and not rely only on the partition.
This is useful for example to update a kernel or a bootloader which is
not part of a partition

It handles the following multiplicative suffixes: K=1024 and M=1024*1024.

This commit has been tested:
* with and without signature
* with and without compression

Signed-off-by: Sid-Ali Teir <steir@witekio.us>
Signed-off-by: Georges Savoundararadj <gsavoundararadj@witekio.us>