network_thread: be sure to not report old messages
The network thread sends IPC messages to other processes and waits for
an answer. If an answer comes too late, the previous message is then
read and reported because it is queued into the socket.

The behavior must be consistent: in case of timeout, a later answer
should be ignored and an error must be forwarded to the caller.

Drop all messages from the IPC socket if any before sending a new
request.

Signed-off-by: Stefano Babic <sbabic@denx.de>