IPC: add a timeout field for answer
IPC to subprocesses is very simple and follows the command / answer
paradigm. Each command is sent and the caller is blocked for an answer.
In case the destination requires much time, this introduce a way to set
the value of the timeout when waiting for an answer.

In case no timeout is put in the IPC message, DEFAULT_INTERNAL_TIMEOUT
is used.

Signed-off-by: Stefano Babic <sbabic@denx.de>