hawkbit: strict behavior for cancelUpdate
The API in case of cancelAction requires to get the URL for cancelAction
during a poll request. If the URL is set, a get to this URL should be
done to retrieve the ID ("stopId") of the action to be confirmed. In all
tests, actionId and stopId are the same, and  the URL for the
cancelAction is always the same.

However, this is not guaranteed to be the same in future and the stopId
could be different as the actionId. Changes are done to be more strict
compliant to Hawkbit's API. The retrieved stopId is used as id to
confirm a cancel request.

ActionId and StopId are now forwarded to external processes via the
progress interface.

Signed-off-by: Stefano Babic <sbabic@denx.de>