downloader: abort slow retry loop after n retries
download_from_url has fast retries (20s) that should catch transient
errors like DNS reconfiguration / router restarts.
The longer retries are intended when the server is in maintenance
or if the connection is currently very slow(GSM). Then we decrease
the network load by taking a longer backoff. This is useful in A/B images
where you want to keep polling the server forever, but limit the
negative impact of increased network load.

Signed-off-by: Andreas Fenkart <afenkart@gmail.com>