LICENSE: relicensing the client library under LGPL
SWUpdate can be driven from an external software via the control socket.
The client library is thought for this purpose and it makes sense that
only this part of the software (the library itself and the required
header) are licensed under Less General Public License (LGPL).
Commercial software can make use of the library without becoming GPL.

Signed-off-by: Stefano Babic <sbabic@denx.de>