Split spawn function for external processes
SWUpdate spawns further processes depending on the configuration. It
does not allow to spawn general command, that could be supervised by
SWUpdate itself.

This feature is covered by systemd, but SWUpdate runs oft on device
where systemd is not installed.

Patch extend the process control mechanism to allow to start a process
passing a command line instead of a pointer to a function.

Signed-off-by: Stefano Babic <sbabic@denx.de>