Spawn processes from configuration file
Read from configuration file if further processes must be started. A
section "processes" (an array) is added to the configuration file and it must have
at least two entries:

	name : this is the name of the process known to SWUpdate
	exec : the command line string to start the process

The process is run as root if not different specified. It is possible to
add a section with the "name" of the process (exactly as it is done for
suricatta / webserver / downloader) that contains "userid" and "groupid"
values.

Signed-off-by: Stefano Babic <sbabic@denx.de>