Add CONFIG_UBIWHITELIST
Some MTD devices are planned to be used with UBI. If an attach fails,
the whole update stops. However, specially in the factory, it is not
guaranteed which is the status of a flash before it is accessed the
first time and it is not always empty. It can happen that a flash has
some data and an attach fails, and SWUpdate stops with error.
After erasing the flash, everything works as expected.

Introduce a CONFIG_UBIWHITELIST to signal MTD devices that have always
UBI. This has a lower priority as CONFIG_UBIBLACKLIST (if a MTD device
is put in both lists, it will be blacklisted). In case UBI attach fails
and the MTD device is whitelisted, erase once the device and tries once
again to call attach.

Signed-off-by: Stefano Babic <sbabic@denx.de>