hawkbit: check for cancellation before verifying an install
If SWUpdate is started in WAIT state, it does not send any feedback to
the Hawkbit server and waits for a message from an external process. If
in the meantime an update is cancelled on the server, the current
install should be not considered as valid anymore. When an external
process asks to close the install, a feedback is sent for an action that
is already cancelled by the server. The Hawkbit server always answers
with HTTP Code = 200 (OK), and the external process can think that the
current update was ACKed and should be run.

Check on the server if there is a cancel URL when a message from the
backchannel IPC and before sending the feedback to the server. A cancel
is then forwarded via progress interface to all registered processes.

Signed-off-by: Stefano Babic <sbabic@denx.de>