u-boot: environment not saved when updating single variable
fw_env_write() just updates the copy of the environment in RAM, and it
does not save it persistently. This is done by fw_env_flush(). The issue
happens when a single variable (as "recovery") is set, while during an
update does not happen because fw_env_parse() is called.

Signed-off-by: Stefano Babic <sbabic@denx.de>