u-boot: fix memory leak when getting variables
botloader_env_get generates memory leak because the internal u-boot
function fw_env_open is called. This function allocates memory for the
environment on the Heap and memory is just freed in the close function.

Because calling fw_env_close() makes the variables not anymore
available, tracks the calls to open to check if the environment was
already copied and can be accessed.

Signed-off-by: Stefano Babic <sbabic@denx.de>