bootloader: change allocation in bootloader_env_get
bootloader_env_get simply returns a pointer to the value of a key inside
the environment. This is notsafe, because a process can alter the
environment corrumping the pointer in use by another process.

Cahnge the behaviour of bootloader_env_get: if a variable is found,
duplicates the string and returns it to the caller.

The caller needs to free the returned string when it does not require it
anymore.

Signed-off-by: Stefano Babic <sbabic@denx.de>
CC: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>