hawkbit: use constant to select IPC mode
Use custom constant to inform get_file that the IPC streaming mode is
requested. While not optimal, this fix at least avoids conficts with the
range of valid file descriptors that starts with value 0.
Value -2 has been chosen to still allow SWUpdate to fail if the
unchecked result of an open (which returns -1 on Linux in case of
errors) is passed.

Signed-off-by: Diego Rondini <diego.rondini@kynetics.com>
Acked-by: Stefano Babic <sbabic@denx.de>