bootloader: race condition when updating u-boot variables
There is a race condition when two processes are updating the
environment at the same time. Each process caches the environment (this
is done automatically by the fw_env_open() function in u-boot library).
When both update the environment, the last one has a old cache without
the changes of the first one, and all setting done by the first
processes are lost.

Change the U-Boot's function and drop the cache. The environment must
always be loaded for any access to the environment. Mutual exclusion is
then guaranteed by the implemented flock() mechanism.

Signed-off-by: Stefano Babic <sbabic@denx.de>