parser: add support for embedded script
It is sometimes required to run some script at the time the
sw-description file is parsed. One typical use case is when the
partitions must be created and one of the artifacts must be streamed
into the partitions. The partitions must be created firstly, and this
cannot happen if the script doing this is packed later in the SWU.

However, there are some other cases when it is required to change the
values passed in sw-description according to run time conditions.

This adds a "hook" to each artifact belonging to "images" or "files",
allowing to run a LUA function embedded in sw-description.

Signed-off-by: Stefano Babic <sbabic@denx.de>